orders/requirements.txt
Django==3.0
pytz==2018.5
djangorestframework~=3.13.1
requests~=2.26.0
PyYAML~=6.0
django-filter~=21.1
psycopg2
celery==5.2.3
redis==4.1.3
pytest~=7.0.0
model-bakery==1.4.0